refactor main to allow cli as high-level lib API
Motivation: use Litestream as a library, and scalable for maintaining multiple
databases use case

Related Issues:
  1. https://github.com/benbjohnson/litestream/issues/440
  2. https://github.com/benbjohnson/litestream/issues/215

Current Workaround: https://github.com/benbjohnson/litestream-library-example

The current workaround is complex, and no guarantee on the API backward
compatibility
(https://github.com/benbjohnson/litestream/issues/215#issuecomment-855426990).

Solution: refactor out main package into `cmd` package

1. Backward Compatibility

CLI interface is more stable than calling Litestream directly via
function calls. Most users already use Litestream through CLI (including
service wrappers), so backward compatibility at this interface level is
desired.

2. Simple

Using Litestream as a library is now as simple as the following:

`cmd.Run([]string{"litestream", "replicate", "/path/to/db", "s3://BUCKETNAME/PATHNAME"})`
`cmd.Run([]string{"litestream", "restore", "-o", "/path/to/db", "s3://BUCKETNAME/PATHNAME"})`
